+++ reqwest-tracing/src/lib.rs
@@ -1,33 +1,35 @@
-//! Opentracing middleware implementation for [`reqwest-middleware`].
+//! Opentracing middleware implementation for [`reqwest_middleware`].
 //!
 //! Attach [`TracingMiddleware`] to your client to automatically trace HTTP requests.
 //!
-//! In this example we define a custom span builder to calculate the request time elapsed and we register the `TracingMiddleware`.
+//! In this example we define a custom span builder to calculate the request time elapsed and we register the [`TracingMiddleware`].
 //! ```rust
-//!
 //! use reqwest_middleware::Result;
 //! use task_local_extensions::Extensions;
 //! use reqwest::{Request, Response};
 //! use reqwest_middleware::ClientBuilder;
-//! use reqwest_tracing::{DefaultRequestOtelSpanBuilder, RequestOtelSpanBuilder, TracingMiddleware};
+//! use reqwest_tracing::{
+//!     impl_on_request_failure, impl_on_request_success, reqwest_otel_span,
+//!     RequestOtelSpanBackend, TracingMiddleware
+//! };
 //! use tracing::Span;
 //! use std::time::{Duration, Instant};
 //!
-//! use reqwest_tracing::root_span;
-//!
 //! pub struct TimeTrace;
 //!
-//! impl RequestOtelSpanBuilder for TimeTrace {
+//! impl RequestOtelSpanBackend for TimeTrace {
 //!     fn on_request_start(req: &Request, extension: &mut Extensions) -> Span {
 //!         extension.insert(Instant::now());
-//!         root_span!(req, time_elapsed = tracing::field::Empty)
+//!         reqwest_otel_span!(req, time_elapsed = tracing::field::Empty)
 //!     }
 //!
 //!     fn on_request_end(span: &Span, outcome: &Result<Response>, extension: &mut Extensions) {
 //!         let time_elapsed = extension.get::<Instant>().unwrap().elapsed().as_millis() as i64;
-//!         DefaultRequestOtelSpanBuilder::on_request_end(span, outcome, extension);
 //!         span.record("time_elapsed", &time_elapsed);
 //!     }
+//!     
+//!     impl_on_request_success!();
+//!     impl_on_request_failure!();
 //! }
 //!
 //! let http = ClientBuilder::new(reqwest::Client::new())
@@ -46,7 +48,10 @@ mod middleware;
 mod otel;
 mod reqwest_otel_span_builder;
 pub use middleware::TracingMiddleware;
-pub use reqwest_otel_span_builder::{DefaultRequestOtelSpanBuilder, RequestOtelSpanBuilder};
+pub use reqwest_otel_span_builder::{
+    default_on_request_failure, default_on_request_success, DefaultSpanBackend,
+    RequestOtelSpanBackend,
+};
 
 #[doc(hidden)]
 pub mod reqwest_otel_span_macro;