+++ reqwest-tracing/src/middleware.rs
@@ -3,37 +3,37 @@ use reqwest_middleware::{Middleware, Next, Result};
 use task_local_extensions::Extensions;
 use tracing::Instrument;
 
-use crate::{DefaultRequestOtelSpanBuilder, RequestOtelSpanBuilder};
+use crate::{DefaultSpanBackend, RequestOtelSpanBackend};
 
 /// Middleware for tracing requests using the current Opentelemetry Context.
-pub struct TracingMiddleware<S: RequestOtelSpanBuilder> {
-    root_span_builder: std::marker::PhantomData<S>,
+pub struct TracingMiddleware<S: RequestOtelSpanBackend> {
+    span_backend: std::marker::PhantomData<S>,
 }
 
-impl<S: RequestOtelSpanBuilder> TracingMiddleware<S> {
+impl<S: RequestOtelSpanBackend> TracingMiddleware<S> {
     pub fn new() -> TracingMiddleware<S> {
         TracingMiddleware {
-            root_span_builder: Default::default(),
+            span_backend: Default::default(),
         }
     }
 }
 
-impl Default for TracingMiddleware<DefaultRequestOtelSpanBuilder> {
+impl Default for TracingMiddleware<DefaultSpanBackend> {
     fn default() -> Self {
         TracingMiddleware::new()
     }
 }
 
-impl<S: RequestOtelSpanBuilder> Clone for TracingMiddleware<S> {
+impl<S: RequestOtelSpanBackend> Clone for TracingMiddleware<S> {
     fn clone(&self) -> Self {
         Self::new()
     }
 }
 
 #[async_trait::async_trait]
-impl<RootSpan> Middleware for TracingMiddleware<RootSpan>
+impl<RequestOtelSpan> Middleware for TracingMiddleware<RequestOtelSpan>
 where
-    RootSpan: RequestOtelSpanBuilder + Sync + Send + 'static,
+    RequestOtelSpan: RequestOtelSpanBackend + Sync + Send + 'static,
 {
     async fn handle(
         &self,
@@ -41,7 +41,7 @@ where
         extensions: &mut Extensions,
         next: Next<'_>,
     ) -> Result<Response> {
-        let request_span = RootSpan::on_request_start(&req, extensions);
+        let request_span = RequestOtelSpan::on_request_start(&req, extensions);
 
         let outcome_future = async {
             // Adds tracing headers to the given request to propagate the OpenTelemetry context to downstream revivers of the request.
@@ -57,7 +57,11 @@ where
 
             // Run the request
             let outcome = next.run(req, extensions).await;
-            RootSpan::on_request_end(&request_span, &outcome, extensions);
+            match &outcome {
+                Ok(res) => RequestOtelSpan::on_request_success(&request_span, res, extensions),
+                Err(err) => RequestOtelSpan::on_request_failure(&request_span, err, extensions),
+            }
+            RequestOtelSpan::on_request_end(&request_span, &outcome, extensions);
             outcome
         };