+++ reqwest-tracing/CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+- Fix how we set the OpenTelemetry span status, based on the HTTP response status.
+
 # [0.2.2] - 2022-04-21
 ### Fixed
 - Opentelemetry context is now propagated when the request span is disabled.